Fixed various %d vs %"PRIrec" issues.
git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@2916 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1

+++ src/gap5/shuffle_pads.c
@@ -1346,7 +1346,7 @@ int remove_pad_columns(GapIO *io, int ncontigs, contig_list_t *contigs,
 		continue;
 
 	    vmessage("  Removing column %d %d%% pad (%d of %d), conf. %f)\n",
-		     j+contigs[i].start,
+		     (int)j+contigs[i].start,
 		     100 * cons[j].counts[4] / cons[j].depth,
 		     cons[j].counts[4], cons[j].depth,
 		     cons[j].scores[cons[j].call]);
